@@ -5,7 +5,7 @@
     "type": "library",
     "license": "BSD-3-Clause",
     "require": {
-        "php": ">=7.0",
+        "php": ">=7.1",
         "psr/log": "^1.0",
         "linio/util": "^2.0",
         "linio/database": "^2.0|^3.0",
@@ -13,9 +13,9 @@
         "doctrine/inflector": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.0",
+        "phpunit/phpunit": ">=5.0",
         "contorion/aerospike-stub": "1.0.*",
-        "friendsofphp/php-cs-fixer": "^2.0"
+        "friendsofphp/php-cs-fixer": "^2.4"
     },
     "suggest": {
         "ext-phpiredis": "Allows faster serialization and deserialization of the Redis protocol",
