@@ -9,7 +9,7 @@ use Linio\Component\Cache\Exception\KeyNotFoundException;
 /**
  * @requires extension aerospike
  */
-class AerospikeAdapterTest extends \PHPUnit_Framework_TestCase
+class AerospikeAdapterTest extends \PHPUnit\Framework\TestCase
 {
     const TEST_NAMESPACE = 'mx';
 
@@ -18,14 +18,14 @@ class AerospikeAdapterTest extends \PHPUnit_Framework_TestCase
      */
     protected $adapter;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->adapter = new AerospikeAdapter(['hosts' => [['addr' => '127.0.0.1', 'port' => 3000]], 'persistent' => true]);
         $this->adapter->setNamespace(static::TEST_NAMESPACE);
         $this->adapter->flush();
     }
 
-    public function testIsSettingAndGetting()
+    public function testIsSettingAndGetting(): void
     {
         $setResult = $this->adapter->set('foo', 'bar');
         $actual = $this->adapter->get('foo');
@@ -34,15 +34,14 @@ class AerospikeAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('bar', $actual);
     }
 
-    /**
-     * @expectedException \Linio\Component\Cache\Exception\KeyNotFoundException
-     */
-    public function testIsGettingInexistentKey()
+    public function testIsGettingInexistentKey(): void
     {
+        $this->expectException(\Linio\Component\Cache\Exception\KeyNotFoundException::class);
+
         $actual = $this->adapter->get('foo');
     }
 
-    public function testIsFindingKey()
+    public function testIsFindingKey(): void
     {
         $this->adapter->set('foo', 'bar');
 
@@ -51,7 +50,7 @@ class AerospikeAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue($actual);
     }
 
-    public function testIsNotFindingKey()
+    public function testIsNotFindingKey(): void
     {
         $this->adapter->setNamespace(static::TEST_NAMESPACE);
         $this->adapter->set('foo', 'bar');
@@ -61,7 +60,7 @@ class AerospikeAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse($actual);
     }
 
-    public function testIsGettingMultipleKeys()
+    public function testIsGettingMultipleKeys(): void
     {
         $this->adapter->setNamespace(static::TEST_NAMESPACE);
         $this->adapter->set('foo', 'bar');
@@ -72,7 +71,7 @@ class AerospikeAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(['foo' => 'bar', 'fooz' => 'baz'], $actual);
     }
 
-    public function testIsGettingMultipleKeysWithInvalidKeys()
+    public function testIsGettingMultipleKeysWithInvalidKeys(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
@@ -82,7 +81,7 @@ class AerospikeAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(['foo' => 'bar'], $actual);
     }
 
-    public function testIsSettingMultipleKeys()
+    public function testIsSettingMultipleKeys(): void
     {
         $actual = $this->adapter->setMulti(['foo' => 'bar', 'fooz' => 'baz']);
 
@@ -91,7 +90,7 @@ class AerospikeAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('baz', $this->adapter->get('fooz'));
     }
 
-    public function testIsDeletingKey()
+    public function testIsDeletingKey(): void
     {
         $this->adapter->set('foo', 'bar');
 
@@ -108,7 +107,7 @@ class AerospikeAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertNull($actual);
     }
 
-    public function testIsDeletingMultipleKeys()
+    public function testIsDeletingMultipleKeys(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
@@ -134,14 +133,14 @@ class AerospikeAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertNull($actual2);
     }
 
-    public function testIsDeletingInexistentKey()
+    public function testIsDeletingInexistentKey(): void
     {
         $actual = $this->adapter->delete('foo');
 
         $this->assertTrue($actual);
     }
 
-    public function testIsDeletingInexistentMultipleKeys()
+    public function testIsDeletingInexistentMultipleKeys(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
@@ -167,7 +166,7 @@ class AerospikeAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('baz', $actual2);
     }
 
-    public function testIsFlushingData()
+    public function testIsFlushingData(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
