@@ -9,7 +9,7 @@ use Linio\Component\Cache\Exception\KeyNotFoundException;
 /**
  * @requires extension apc
  */
-class ApcAdapterTest extends \PHPUnit_Framework_TestCase
+class ApcAdapterTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * @var ApcAdapter
@@ -21,7 +21,7 @@ class ApcAdapterTest extends \PHPUnit_Framework_TestCase
      */
     protected $namespace;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->adapter = new ApcAdapter();
         $this->namespace = 'mx';
@@ -29,7 +29,7 @@ class ApcAdapterTest extends \PHPUnit_Framework_TestCase
         apc_clear_cache('user');
     }
 
-    public function testIsSettingAndGetting()
+    public function testIsSettingAndGetting(): void
     {
         $setResult = $this->adapter->set('foo', 'bar');
         $actual = $this->adapter->get('foo');
@@ -38,15 +38,14 @@ class ApcAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('bar', $actual);
     }
 
-    /**
-     * @expectedException \Linio\Component\Cache\Exception\KeyNotFoundException
-     */
-    public function testIsGettingInexistentKey()
+    public function testIsGettingInexistentKey(): void
     {
+        $this->expectException(\Linio\Component\Cache\Exception\KeyNotFoundException::class);
+
         $actual = $this->adapter->get('foo');
     }
 
-    public function testIsFindingKey()
+    public function testIsFindingKey(): void
     {
         $this->adapter->set('foo', 'bar');
 
@@ -55,7 +54,7 @@ class ApcAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue($actual);
     }
 
-    public function testIsNotFindingKey()
+    public function testIsNotFindingKey(): void
     {
         $this->adapter->set('foo', 'bar');
 
@@ -64,7 +63,7 @@ class ApcAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse($actual);
     }
 
-    public function testIsGettingMultipleKeys()
+    public function testIsGettingMultipleKeys(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
@@ -74,7 +73,7 @@ class ApcAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(['foo' => 'bar', 'fooz' => 'baz'], $actual);
     }
 
-    public function testIsGettingMultipleKeysWithInvalidKeys()
+    public function testIsGettingMultipleKeysWithInvalidKeys(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
@@ -84,7 +83,7 @@ class ApcAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(['foo' => 'bar'], $actual);
     }
 
-    public function testIsSettingMultipleKeys()
+    public function testIsSettingMultipleKeys(): void
     {
         $actual = $this->adapter->setMulti(['foo' => 'bar', 'fooz' => 'baz']);
 
@@ -93,7 +92,7 @@ class ApcAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('baz', $this->adapter->get('fooz'));
     }
 
-    public function testIsDeletingKey()
+    public function testIsDeletingKey(): void
     {
         $this->adapter->set('foo', 'bar');
 
@@ -110,7 +109,7 @@ class ApcAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertNull($actual);
     }
 
-    public function testIsDeletingMultipleKeys()
+    public function testIsDeletingMultipleKeys(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
@@ -136,14 +135,14 @@ class ApcAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertNull($actual2);
     }
 
-    public function testIsDeletingInexistentKey()
+    public function testIsDeletingInexistentKey(): void
     {
         $actual = $this->adapter->delete('foo');
 
         $this->assertTrue($actual);
     }
 
-    public function testIsDeletingInexistentMultipleKeys()
+    public function testIsDeletingInexistentMultipleKeys(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
@@ -169,7 +168,7 @@ class ApcAdapterTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('baz', $actual2);
     }
 
-    public function testIsFlushingData()
+    public function testIsFlushingData(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
