@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace Linio\Component\Cache\Adapter;
 
 use Linio\Component\Cache\Exception\KeyNotFoundException;
-use PHPUnit_Framework_Assert;
+use PHPUnit\Framework\Assert;
 
 /**
  * @requires extension memcached
  */
-class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
+class MemcachedAdapterPersistentTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * @var ApcAdapter
@@ -22,7 +22,7 @@ class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
      */
     protected $namespace;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->adapter = new MemcachedAdapter($this->getMemcachedPersistentTestConfiguration());
         $this->namespace = 'mx';
@@ -30,17 +30,17 @@ class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
         $this->adapter->flush();
     }
 
-    public function testIsCreatingPersistentConnection()
+    public function testIsCreatingPersistentConnection(): void
     {
         /** @var $client \Memcached */
-        $client = PHPUnit_Framework_Assert::readAttribute($this->adapter, 'memcached');
+        $client = Assert::readAttribute($this->adapter, 'memcached');
         $this->assertTrue($client->isPersistent());
     }
 
-    public function testIsRespectingPoolSize()
+    public function testIsRespectingPoolSize(): void
     {
         $connection = new MemcachedAdapter($this->getMemcachedPersistentTestConfiguration());
-        $client1 = PHPUnit_Framework_Assert::readAttribute($connection, 'memcached');
+        $client1 = Assert::readAttribute($connection, 'memcached');
         /** @var $client1 \Memcached */
         $stats = $client1->getStats();
         $serverStats = reset($stats);
@@ -52,7 +52,7 @@ class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
             $connections[] = $connection;
         }
 
-        $client100 = PHPUnit_Framework_Assert::readAttribute($connection, 'memcached');
+        $client100 = Assert::readAttribute($connection, 'memcached');
         /** @var $client100 \Memcached */
         $stats = $client100->getStats();
         $serverStats = reset($stats);
@@ -61,7 +61,7 @@ class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
         $this->assertLessThanOrEqual(15, $currentConnections - $initialConnections);
     }
 
-    public function testIsSettingAndGetting()
+    public function testIsSettingAndGetting(): void
     {
         $setResult = $this->adapter->set('foo', 'bar');
         $actual = $this->adapter->get('foo');
@@ -70,15 +70,14 @@ class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('bar', $actual);
     }
 
-    /**
-     * @expectedException \Linio\Component\Cache\Exception\KeyNotFoundException
-     */
-    public function testIsGettingInexistentKey()
+    public function testIsGettingInexistentKey(): void
     {
+        $this->expectException(\Linio\Component\Cache\Exception\KeyNotFoundException::class);
+
         $actual = $this->adapter->get('foo');
     }
 
-    public function testIsFindingKey()
+    public function testIsFindingKey(): void
     {
         $this->adapter->set('foo', 'bar');
 
@@ -87,7 +86,7 @@ class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue($actual);
     }
 
-    public function testIsNotFindingKey()
+    public function testIsNotFindingKey(): void
     {
         $this->adapter->set('foo', 'bar');
 
@@ -96,7 +95,7 @@ class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse($actual);
     }
 
-    public function testIsGettingMultipleKeys()
+    public function testIsGettingMultipleKeys(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
@@ -106,7 +105,7 @@ class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(['foo' => 'bar', 'fooz' => 'baz'], $actual);
     }
 
-    public function testIsGettingMultipleKeysWithInvalidKeys()
+    public function testIsGettingMultipleKeysWithInvalidKeys(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
@@ -116,7 +115,7 @@ class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(['foo' => 'bar'], $actual);
     }
 
-    public function testIsSettingMultipleKeys()
+    public function testIsSettingMultipleKeys(): void
     {
         $actual = $this->adapter->setMulti(['foo' => 'bar', 'fooz' => 'baz']);
 
@@ -125,7 +124,7 @@ class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('baz', $this->adapter->get('fooz'));
     }
 
-    public function testIsDeletingKey()
+    public function testIsDeletingKey(): void
     {
         $this->adapter->set('foo', 'bar');
 
@@ -142,7 +141,7 @@ class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
         $this->assertNull($actual);
     }
 
-    public function testIsDeletingMultipleKeys()
+    public function testIsDeletingMultipleKeys(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
@@ -168,14 +167,14 @@ class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
         $this->assertNull($actual2);
     }
 
-    public function testIsDeletingInexistentKey()
+    public function testIsDeletingInexistentKey(): void
     {
         $actual = $this->adapter->delete('foo');
 
         $this->assertTrue($actual);
     }
 
-    public function testIsDeletingInexistentMultipleKeys()
+    public function testIsDeletingInexistentMultipleKeys(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
@@ -201,7 +200,7 @@ class MemcachedAdapterPersistentTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('baz', $actual2);
     }
 
-    public function testIsFlushingData()
+    public function testIsFlushingData(): void
     {
         $this->adapter->set('foo', 'bar');
         $this->adapter->set('fooz', 'baz');
